requirements.txt
matplotlib>=3.3.4
numpy>=1.22.0
Pillow>=9.2.0
tifffile>=2020.10.1
timm>=0.4.12
torch>=1.10.2
torchsummary>=1.5.1
torchvision>=0.11.3
tqdm>=4.63.0
